backend/requirements.txt
# Core Python dependencies for BiblioChat Backend

# AI & Embeddings
cohere>=5.0.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Data Validation
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Vector Database
qdrant-client>=1.7.0

# Environment Variables
python-dotenv>=1.0.0